Calculate Economic Impacts

Apply the multipliers and Leontief inverse parameters to the change in final demand vector in order to model the impact of changes in demand on output, gross value added, and employment. Currently this function produces type 0 (direct), type 1 (indirect), and type 2 (induced) effects on output, GVA, and employment.

EconEffectsCalc(
  leontief,
  fdemand,
  FAI = TRUE,
  year = 2019,
  tax_data = tobalciomodel::inctax_params
)

Arguments

leontief
List object. The output of LeonfiefCalc containing multipliers and leontief matrices.

fdemand
Data table. The output of PrepFinalDemand.

FAI
Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the ONS ones. Defaults to FALSE.

year
Integer. Select the year (from 2016 to 2020) of alcohol, tobacco, earnings and employment data to use in the analysis.

tax_data
Data table. Package data containing the necessary parameters to calculate income tax and national insurance contributions from annual earnings data

Value

A data table of economic impacts by sector

Examples

if (FALSE) {

}