Generate IO Inputs from STAPM Models

Function to read in and further process the outputs of STAPM price policy modelling economic outcomes. Produces the change in demand for tobacco, on-trade alcohol and off-trade alcohol measured in basic prices and also the net change in government revenue.

GenInputsSTAPM(
  root = NULL,
  label = NULL,
  label2 = NULL,
  min_age = NULL,
  index_year = NULL,
  policy_effect_year = NULL,
  n_years = 1,
  write_inputs,
  name
)

Arguments

root
Character - the file path to the folder storing STAPM raw output

label
Character - the label used to identify the set of model runs.

label2
Character vector - vector of treatment labels identifying specific model runs.

min_age
Numeric - Youngest age included in the model.

index_year
Numeric - Model index year.

policy_effect_year
Numeric - Year in which new price policy is implemented in STAPM.

n_years
Numeric - Number of years including the policy effect year for which to generate inputs (defaults to 1 - policy effect year only).

write_inputs
Character - file path to the folder to which the input data will be written.

name
Character - file name for the input data.

Value

Examples

if (FALSE) {

}