Read IO Model Inputs from STAPM Output

Function to process the outputs of STAPM price policy modelling economic outcomes. Produces the change in demand for tobacco, on-trade alcohol and off-trade alcohol measured in basic prices and also the net change in government revenue.

ReadInputs(data, policy_effect_year, n_years = 1, govt = TRUE)

Arguments

data
Data table. Output from the `stapmr` package function `EconCalc()`

policy_effect_year
Numeric. Year for which economic impacts are to be modelled.

n_years
Numeric. Number of years starting with the policy effect year for which to extract treatment effects.

govt
Logical. If TRUE include changes in government revenue in the analysis.

Value

Examples

if (FALSE) {

### Read in the output of a price policy simulation

### Obtain economic outcomes

### use the output from EconCalc as the input to the
### function to clean IO model inputs.

iomodel_inputs <- ReadInputs(econ_calcs)

}