Reallocate Household Expenditure

A function to reallocate, based on assumptions about how individuals redistribute spending, the expenditure saved as a result of a policy or exogenous changes in consumer preferences.

ReallocateHhold(
  expenditure = c(-20, -10, -30),
  hhold_passthru = 1,
  vector = 3,
  vectors_data = tobalciomodel::vectors_hhold,
  mapping = tobalciomodel::coicop_cpa_mapping,
  FAI = T
)

Arguments

expenditure
Numeric vector. Change in household consumption measured in basic prices for off-trade alcohol, on-trade alcohol, and tobacco.

hhold_passthru
Numeric. Assumed household rate of passthrough - the proportion of change in spending which is compensated for in spending on other consumption categories. Defaults to 1 (full passthrough).

vector
Numeric (1-3). The distribution of reallocation of spending to implement from the vectors_hhold data. Option 1 allocates pro-rata across all consumption categories, option 2 excludes alcohol and tobacco consumption, option 3 (default) further excludes health, education, rents and utilities.

mapping
data table containing the mapping algorithm from COICOP to CPA.

FAI
Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the ONS ones. Defaults to FALSE.

vector_data
data table containing the redistribution vectors.